+++ pkg/exchange_client/kucoin/errors.go
@@ -15,6 +15,9 @@ func wrapKuCoinError(msg string, code int) error {
 	if strings.Contains(msgLower, "locked from withdrawal") {
 		return fmt.Errorf("kucoin error: %s (%d): %w", msg, code, exchangeclient.ErrWithdrawalBalanceLocked)
 	}
+	if strings.Contains(msgLower, "funds were still awaiting confirmation") {
+		return fmt.Errorf("kucoin errors: %s (%d): %w", msg, code, exchangeclient.ErrWithdrawalBalanceLocked)
+	}
 
 	return fmt.Errorf("kucoin error: %s (%d)", msg, code)
 }